% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wiki.R
\name{clean_wiki}
\alias{clean_wiki}
\title{Clean Your Wiki}
\usage{
clean_wiki(string)
}
\arguments{
\item{string}{The string to be cleaned}
}
\description{
This function accepts a string and removes new line indentations and html tags with regex
}
\examples{
clean_wiki("<p>some text</p>")
}
