% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wiki.R
\name{get_wiki}
\alias{get_wiki}
\title{Get the Text of a Wikipedia Article}
\usage{
get_wiki(title, clean = TRUE)
}
\arguments{
\item{title}{The title or (titles) of the Wikipedia page to be searched. If you would like to query multiple articles,
put the titles in a character vector}

\item{clean}{Should getwiki remove html tags from the returned text?}
}
\description{
Get the text of a wikipedia article by submitted a title as a string.
}
\examples{
get_wiki("United States")
get_wiki(c("United States", "France"))
}
