% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_wiki.R
\name{random_wiki}
\alias{random_wiki}
\title{Get the Text of a Random Wikipedia Article}
\usage{
random_wiki(clean = TRUE)
}
\arguments{
\item{clean}{Should getwiki remove html tags from the returned text?}
}
\value{
\code{random_wiki} will return a single named character value whose value is the text of the wikipedia page
}
\description{
Get the text of a random wikipedia article
}
\examples{
random_wiki()
}
