% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_wiki.R
\name{search_wiki}
\alias{search_wiki}
\title{Search Wikipedia for Articles}
\usage{
search_wiki(search_term, clean = TRUE)
}
\arguments{
\item{search_term}{The search term you would like to use.}

\item{clean}{Should getwiki remove html tags from the returned text?}
}
\description{
Get the text of a wikipedia article by submitted a title as a string.
}
\examples{
search_wiki("Belgrade")
}
