% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_wiki.R
\name{trend_wiki}
\alias{trend_wiki}
\title{Find the Page Views for an Article}
\usage{
trend_wiki(title)
}
\arguments{
\item{title}{The title of the Wikipedia article you would like trends for.}
}
\value{
trend_wiki will return a dataframe of the past sixty days of page views for the requested title
}
\description{
Find the page views for an article for the past sixty days.
}
\examples{
trend_wiki("Belgrade")
}
