% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_wiki.R
\name{search_wiki}
\alias{search_wiki}
\title{Search Wikipedia for Articles}
\usage{
search_wiki(search_term, clean = TRUE)
}
\arguments{
\item{search_term}{The search term you would like to use.}

\item{clean}{Should getwiki remove html tags from the returned text?}
}
\value{
search_term will return a dataframe of the top twenty search results. The "title" column will be the title
of the articles and the "content" column will be the first paragraph from those articles.
}
\description{
Search for the top twenty wikipedia pages that match a given query. This function will return a dataframe with the
names of the matched articles and the first paragraph of content.
}
\examples{
search_wiki("Belgrade")
}
